* install alpine.js and change the create customer post to htmx and remove some cruft

* demonstrate successful implementation of alpine. It. Is. Hideous.

* add try catch to customer creation to try to track down why there is a re-render within the fragment maybe?

* squashMe freeze a failed strategy

* restore original functionality

* clean cruft; continue troubleshooting

* Basic Customer/Contract view

Adds contract and service model.
Adds Contract and Customer factory.
Adds new layout with Customer and Single Customer View.
Adds editing customers.
Adds service table.

* Basic Contract View

Adds contract view.

* Update CustomerController

Simplifies web routes with CustomerController.

* Add ContractController, renamed Contract pages

* Added ContactFactory, created Layout for Customer Form

* Install alpine and htmx, use the cdn version of alpine to get correct behavior, auto-format resources/views/customers/index.blade.php, add alpine handling of creation form showing. Remove now-crufty toggleCreateForm method

* Adds Contract/Service Controllers

Adds Services view.
Adds ability to add contract.

* adds apline functionality to the cancel button

* remove alpine from bootstrap

* get the partial fragment render to work for the customer creation

* implement delete using htmx

* remove customer-management.blade.php and remove cruft

* implement table-layout everywhere

* remove crufty method

* remove more cruft and auto-format some stuff

* get the toast to work as expected at the expense of the htmx rendering

* recover functionality of fragment return of customer create

* continue troubleshooting (and breaking) the form errors issue

* add functionality of the deletion indicator

* make some fixes to darbi_account

* rename darbi_account to darbi_customer_account_number globally

* make last changes with Greg and Skylar. Note that this may not function currently and definitely needs to be refactored

* Remove excess routes in web.php

* get create and update to work from the same blade template

* fix bug where cancel edit form was submitting the form

* respond to self code-review and remove cruft

* Added customer routes back + removed Symbiota image source

* Add contract_service table

Pivot table for contracts and services.
Add service factory.
Changed Service id to unsigned int instead of unsigned big int.

* get validation with htmx to work

* Show services variables on contract

* Automatic customer id in contract creation

If user creates contract from customer page, customer ID is automatically filled in.

* fix bug where clearing after an error did not clear the error list

* resolve error handling issue by just making the conditional error div section part of the customer-list-div div and fragment

* Ability to add multiple services to a contract

Adds service selection in contract.create
Adds contact view

* Misc. Contract text

Made contract.show neater

* remove cruft

* remove uneccesary method for toast, and implement form reset upon successful customer creation

* remove even more cruft

* Blade layouts for input boxes

Adds blade layouts for input boxes for easier use

* Basic contract edit system

Plus using layouts for customer-form

* Added basic User Registration

Form validation for user registration and log in, and action to log out.
Adds authentication to routes, pages can only be seen when user is logged in.
Adds "old('value')" to forms so previous error submissions show.
Adds "exists" requirements for creating contracts.

* Added creating and editing services

* Add basic retire function to services

Adds new variable 'active_status' to service.
User can retire active services.
Also adds database seeder.

* Minor touchups

* fix some DARBI capitalization issues and add new cascade behavior to contracts when their parent customers are deleted, thereby restoring successful deletion

* Darbi->DARBI and misc. changes

Changed Darbi to DARBI.
Adds price-per-unit to service.index view.
Some reformatting.

---------

+++ routes/web.php
@@ -1,13 +1,46 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
-use App\Models\Customer;
 use App\Http\Controllers\CustomerController;
+use App\Http\Controllers\ContractController;
+use App\Http\Controllers\ServiceController;
+use App\Http\Controllers\RegisteredUserController;
+use App\Http\Controllers\SessionController;
+use App\Models\Contact;
 
 Route::get('/', function () {
     return view('welcome');
 });
-Route::get('/customer-management', [CustomerController::class, 'index']);
-Route::post('/customer', [CustomerController::class, 'create']);
-Route::delete('/customer/{id}', [CustomerController::class, 'destroy']);
+
+Route::post('/customer', [CustomerController::class, 'create'])->middleware('auth');
+Route::delete('/customer/{id}', [CustomerController::class, 'destroy'])->middleware('auth');
+
+Route::get('/services', [ServiceController::class, 'index'])->middleware('auth');
+Route::get('/services/{service}', [ServiceController::class, 'show'])->middleware('auth');
+Route::patch('/services/{service}', [ServiceController::class, 'update'])->middleware('auth');
+Route::post('/services', [ServiceController::class, 'store'])->middleware('auth');
+Route::patch('/services/{service}/retire', [ServiceController::class, 'retire'])->middleware('auth');
+
+Route::get('/contracts/create/{customer?}', [ContractController::class, 'create'])->name('contracts.create')->middleware('auth');
+Route::get('/contracts', [ContractController::class, 'index'])->middleware('auth');
+Route::get('/contracts/{contract}', [ContractController::class, 'show'])->middleware('auth');
+Route::post('/contracts', [ContractController::class, 'store'])->middleware('auth');
+Route::patch('/contracts/{contract}', [ContractController::class, 'update'])->middleware('auth');
+
+Route::get('/customers', [CustomerController::class, 'index'])->middleware('auth');
+Route::get('/customers/{customer}', [CustomerController::class, 'show'])->middleware('auth');
+Route::post('/customers', [CustomerController::class, 'create'])->middleware('auth');
+Route::patch('/customers/{customer}', [CustomerController::class, 'update'])->middleware('auth');
+
+Route::get('/contacts', function (){
+    return view('contacts.index', ['contacts' => Contact::all()]);
+})->middleware('auth');
+
+Route::get('/register', [RegisteredUserController::class, 'create']);
+Route::post('/register', [RegisteredUserController::class, 'store']);
+
+Route::get('/login', [SessionController::class, 'create']);
+Route::post('/login', [SessionController::class, 'store']);
+Route::post('/logout', [SessionController::class, 'destroy']);
+
 ?>
